add Repeat extension method to SuitRankBuilder class
Make SuitRankBuilder class static

This commit introduces the Repeat extension method to the SuitRankBuilder class. The Repeat method takes an array of SuitRankCard objects and an optional integer parameter 'times', which specifies the number of times the array should be repeated. If 'times' is less than 1, an ArgumentException is thrown.

The Repeat method creates a new array with a length equal to the length of the input array multiplied by the 'times' parameter. It then copies the elements of the input array 'times' times into the new array, ensuring that the original order of elements is preserved.